refactor: unify message and exception creation order in unit tests
