+++ src/ExceptionalMatcher/Rule/Object/Property/Match/Condition/Value/Tests/ExceptionValueMatchConditionUnitTest.php
@@ -50,11 +50,12 @@ final class ExceptionValueMatchConditionUnitTest extends TestCase
 
     public function testValueExceptionCondition(): void
     {
+        $message = HandleableMessageStub::create();
+
         $exceptionAdapter = new CompositeException([
             new SomeValueException('matched!'),
             new SomeValueException('whatever'),
         ]);
-        $message = HandleableMessageStub::create();
 
         $matchedExceptionList = $this->matcher->match($exceptionAdapter, $message);
